+++ src/pharmpy/tools/__init__.py
@@ -6,6 +6,7 @@ __all__ = (
     'predict_influential_individuals',  # pyright: ignore [reportUnsupportedDunderAll]
     'predict_influential_outliers',  # pyright: ignore [reportUnsupportedDunderAll]
     'predict_outliers',  # pyright: ignore [reportUnsupportedDunderAll]
+    'print_fit_summary',  # pyright: ignore [reportUnsupportedDunderAll]
     'read_results',  # pyright: ignore [reportUnsupportedDunderAll]
     'retrieve_final_model',  # pyright: ignore [reportUnsupportedDunderAll]
     'retrieve_models',  # pyright: ignore [reportUnsupportedDunderAll]
@@ -27,6 +28,7 @@ _allowed = set(__all__)
 _run_keys = {
     'create_results',
     'fit',
+    'print_fit_summary',
     'read_results',
     'retrieve_final_model',
     'retrieve_models',